pyproject.toml
[project]
name = "ai-group-manager"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Snowykami", email = "snowykami@outlook.com"},
]
dependencies = ["nonebot2>=2.4.1", "nonebot-plugin-marshoai>=1.0.7"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
